+++ lib/src/tinycolor.dart
@@ -1,11 +1,10 @@
 import 'dart:math' as math;
 import 'dart:ui' show Color;
 
-import 'package:flutter/painting.dart' show HSVColor;
+import 'package:flutter/painting.dart' show HSLColor, HSVColor;
 import 'package:pigment/pigment.dart';
 
 import 'conversion.dart';
-import 'hsl_color.dart';
 import 'util.dart';
 
 class TinyColor {
@@ -23,7 +22,7 @@ class TinyColor {
   }) =>
       TinyColor(Color.fromARGB(a, r, g, b));
 
-  factory TinyColor.fromHSL(HslColor hsl) => TinyColor(hslToColor(hsl));
+  factory TinyColor.fromHSL(HSLColor hsl) => TinyColor(hsl.toColor());
 
   factory TinyColor.fromHSV(HSVColor hsv) => TinyColor(hsv.toColor());
 
@@ -51,19 +50,7 @@ class TinyColor {
 
   HSVColor toHsv() => HSVColor.fromColor(_color);
 
-  HslColor toHsl() {
-    final hsl = rgbToHsl(
-      r: _color.red.toDouble(),
-      g: _color.green.toDouble(),
-      b: _color.blue.toDouble(),
-    );
-    return HslColor(
-      h: hsl.h * 360,
-      s: hsl.s,
-      l: hsl.l,
-      a: _color.alpha.toDouble(),
-    );
-  }
+  HSLColor toHsl() => HSLColor.fromColor(_color);
 
   String toHex8() => _color.value.toRadixString(16).padLeft(8, '0');
 
@@ -71,9 +58,8 @@ class TinyColor {
 
   TinyColor lighten([int amount = 10]) {
     final hsl = toHsl();
-    hsl.l += amount / 100;
-    hsl.l = clamp01(hsl.l);
-    return TinyColor.fromHSL(hsl);
+    final lightness = clamp01(hsl.lightness + amount / 100);
+    return TinyColor.fromHSL(hsl.withLightness(lightness));
   }
 
   TinyColor brighten([int amount = 10]) {
@@ -106,9 +92,8 @@ class TinyColor {
 
   TinyColor darken([int amount = 10]) {
     final hsl = toHsl();
-    hsl.l -= amount / 100;
-    hsl.l = clamp01(hsl.l);
-    return TinyColor.fromHSL(hsl);
+    final lightness = clamp01(hsl.lightness - amount / 100);
+    return TinyColor.fromHSL(hsl.withLightness(lightness));
   }
 
   TinyColor tint([int amount = 10]) => mix(
@@ -123,25 +108,22 @@ class TinyColor {
 
   TinyColor desaturate([int amount = 10]) {
     final hsl = toHsl();
-    hsl.s -= amount / 100;
-    hsl.s = clamp01(hsl.s);
-    return TinyColor.fromHSL(hsl);
+    final saturation = clamp01(hsl.saturation - amount / 100);
+    return TinyColor.fromHSL(hsl.withSaturation(saturation));
   }
 
   TinyColor saturate([int amount = 10]) {
     final hsl = toHsl();
-    hsl.s += amount / 100;
-    hsl.s = clamp01(hsl.s);
-    return TinyColor.fromHSL(hsl);
+    final saturation = clamp01(hsl.saturation + amount / 100);
+    return TinyColor.fromHSL(hsl.withSaturation(saturation));
   }
 
   TinyColor greyscale() => desaturate(100);
 
   TinyColor spin(double amount) {
     final hsl = toHsl();
-    final hue = (hsl.h + amount) % 360;
-    hsl.h = hue < 0 ? 360 + hue : hue;
-    return TinyColor.fromHSL(hsl);
+    final hue = (hsl.hue + amount) % 360;
+    return TinyColor.fromHSL(hsl.withHue(hue < 0 ? 360 + hue : hue));
   }
 
   TinyColor mix(Color toColor, [int amount = 50]) {
@@ -157,8 +139,8 @@ class TinyColor {
 
   TinyColor complement() {
     final hsl = toHsl();
-    hsl.h = (hsl.h + 180) % 360;
-    return TinyColor.fromHSL(hsl);
+    final hue = (hsl.hue + 180) % 360;
+    return TinyColor.fromHSL(hsl.withHue(hue));
   }
 
   Color get color => _color;